Move code for isobot New Years special in-game event to archive
