+++ main.py
@@ -109,19 +109,6 @@ with open("themes/halloween.theme.json", 'r') as f:
         print(f"{colors.red}The theme file being loaded might be broken. Rolling back to default configuration...{colors.end}")
         color = discord.Color.random()
 
-# Discord UI Views
-class PresentsDrop(discord.ui.View):
-    @discord.ui.button(label="🎁 Collect Presents", style=discord.ButtonStyle.blurple)
-    async def receive(self, button: discord.ui.Button, interaction: discord.Interaction):
-        presents_bounty = randint(500, 1000)
-        presents[str(interaction.user.id)] += presents_bounty
-        save()
-        button.disabled = True
-        button.label = f"Presents Collected!"
-        button.style = discord.ButtonStyle.grey
-        newembed = discord.Embed(description=f"{interaction.user.name} collected **{presents_bounty} :gift: presents** from this drop!", color=discord.Color.green())
-        await interaction.response.edit_message(embed=newembed, view=self)
-
 #Events
 @client.event
 async def on_ready():
@@ -139,14 +126,6 @@ __________________________________________________""")
     print("[main/FLASK] Starting pinger service...")
     utils.ping.host()
     time.sleep(5)
-    print("[main/Presents] Presents daemon started.")
-    while True:
-        print(f"[main/Presents] Dropping new presents in {colors.cyan}#general{colors.end} channel...")
-        await asyncio.sleep(randint(450, 600))
-        if floor(time.time()) > 1672511400:
-            cyberspace_channel_context = await client.fetch_channel(880409977074888718)
-            localembed = discord.Embed(title="**:gift: Presents have dropped in chat!**", description="Be the first one to collect them!", color=color)
-            await cyberspace_channel_context.send(embed=localembed, view=PresentsDrop()) 
 
 @client.event
 async def on_message(ctx):
